+++ Doctrine/Phpcr/BaseBestMatcher.php
@@ -1,0 +1,58 @@
+<?php
+
+/*
+ * This file is part of the Symfony CMF package.
+ *
+ * (c) 2011-2014 Symfony CMF
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Cmf\Bundle\SeoBundle\Doctrine\Phpcr;
+
+use Doctrine\Common\Persistence\ManagerRegistry;
+use Doctrine\ODM\PHPCR\DocumentManager;
+use Symfony\Cmf\Bundle\SeoBundle\BestMatcherInterface;
+
+/**
+ * Abstract BestMatcher for those who needs make use of base
+ * phpcr-odm classes.
+ *
+ * @author Maximilian Berghoff <Maximilian.Berghoff@gmx.de>
+ */
+abstract class BaseBestMatcher implements BestMatcherInterface
+{
+    /**
+     * @var ManagerRegistry
+     */
+    protected $managerRegistry;
+
+    /**
+     * By concatenating the routeBasePath and the url
+     * we will get the absolute path a route document
+     * should be persisted with.
+     *
+     * @var string
+     */
+    protected $routeBasePath;
+
+    /**
+     * @param ManagerRegistry $managerRegistry
+     * @param $routeBasePath
+     */
+    public function __construct(ManagerRegistry $managerRegistry, $routeBasePath)
+    {
+        $this->managerRegistry = $managerRegistry;
+        $this->routeBasePath = $routeBasePath;
+    }
+
+    /**
+     * @param string                $class
+     * @return null|DocumentManager
+     */
+    public function getManagerForClass($class)
+    {
+        return $this->managerRegistry->getManagerForClass($class);
+    }
+}