+++ Tests/WebTest/Admin/SeoContentAdminExtensionTest.php
@@ -69,7 +69,7 @@ class SeoContentAdminExtensionTest extends BaseTestCase
         $res = $this->client->getResponse();
 
         $this->assertEquals(200, $res->getStatusCode());
-        $this->assertCount(1, $crawler->filter('html:contains("Key")'));
+            $this->assertCount(1, $crawler->filter('html:contains("Key")'));
         $this->assertCount(1, $crawler->filter('html:contains("Value")'));
     }
 }