+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 Changelog
 =========
 
+<<<<<<< HEAD
 * **2014-09-22**: Added the `SeoAwareTrait` to ease creation of SeoAware classes
+=======
+>>>>>>> origin/custom-exception-or-errors
+* **2014-08-02**: Custom exception controller for error handling
 
 1.1.0-RC1
 ---------